[LIVY-705][THRIFT] Refactor common code into a method in the parent class.
